backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
PyYAML
python-dotenv
geopandas
shapely
pyogrio
rtree
numpy
pandas
h3
